requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
DateTime==4.4
idna==3.3
pytz==2021.3
requests==2.27.1
urllib3==1.26.8
zope.interface==5.4.0

python-dotenv~=0.19.2